+++ src/components/navigate.tsx
@@ -1,20 +1,20 @@
 import React, { useEffect } from 'react';
 import { useChangeScreen } from '@common/hooks/use-change-screen';
-import { ScreenPaths } from '@common/types';
+import { RouteUrls } from '@routes/route-urls';
 import { useNavigate, ToOptions } from 'react-router-dom';
 
 interface NavigateProps {
   to: string | ToOptions;
-  screenPath: ScreenPaths;
+  screenPath: RouteUrls;
 }
 
 export const Navigate: React.FC<NavigateProps> = ({ to, screenPath }) => {
-  const doChangeScreen = useChangeScreen();
+  const changeScreen = useChangeScreen();
   const navigate = useNavigate();
   useEffect(() => {
     navigate(to);
-    doChangeScreen(screenPath, false);
-  }, [screenPath, doChangeScreen, to, navigate]);
+    changeScreen(screenPath, false);
+  }, [screenPath, changeScreen, to, navigate]);
 
   return null;
 };