+++ src/features/network-drawer/networks-drawer.tsx
@@ -2,7 +2,7 @@ import React, { useCallback } from 'react';
 import { Box, Button } from '@stacks/ui';
 import { ControlledDrawer } from '@components/drawer/controlled';
 import { useChangeScreen } from '@common/hooks/use-change-screen';
-import { ScreenPaths } from '@common/types';
+import { RouteUrls } from '@routes/route-urls';
 import { useDrawers } from '@common/hooks/use-drawers';
 import { useShowNetworksStore } from '@store/ui/ui.hooks';
 import { NetworkList } from '@features/network-drawer/network-list';
@@ -11,14 +11,14 @@ import { useAnalytics } from '@common/hooks/analytics/use-analytics';
 export const NetworksDrawer: React.FC = () => {
   const { setShowNetworks } = useDrawers();
   const [isShowing] = useShowNetworksStore();
-  const doChangeScreen = useChangeScreen();
+  const changeScreen = useChangeScreen();
   const analytics = useAnalytics();
 
   const handleAddNetworkClick = useCallback(() => {
     void analytics.track('add_network');
     setShowNetworks(false);
-    doChangeScreen(ScreenPaths.ADD_NETWORK);
-  }, [analytics, setShowNetworks, doChangeScreen]);
+    changeScreen(RouteUrls.AddNetwork);
+  }, [analytics, setShowNetworks, changeScreen]);
 
   return (
     <ControlledDrawer