+++ src/pages/choose-account/choose-account.tsx
@@ -2,7 +2,7 @@ import React, { memo, useCallback, useEffect } from 'react';
 import { Title } from '@components/typography';
 import { Accounts } from '@pages/choose-account/components/accounts';
 import { AppIcon } from '@components/app-icon';
-import { ScreenPaths } from '@common/types';
+import { RouteUrls } from '@routes/route-urls';
 import { useWallet } from '@common/hooks/use-wallet';
 import { Navigate } from '@components/navigate';
 import { Header } from '@components/header';
@@ -15,11 +15,11 @@ interface ChooseAccountProps {
 
 export const ChooseAccount: React.FC<ChooseAccountProps> = memo(() => {
   const { name: appName } = useAppDetails();
-  const { wallet, handleCancelAuthentication } = useWallet();
+  const { wallet, cancelAuthentication } = useWallet();
 
   const handleUnmount = useCallback(async () => {
-    handleCancelAuthentication();
-  }, [handleCancelAuthentication]);
+    cancelAuthentication();
+  }, [cancelAuthentication]);
 
   useEffect(() => {
     window.addEventListener('beforeunload', handleUnmount);
@@ -27,7 +27,7 @@ export const ChooseAccount: React.FC<ChooseAccountProps> = memo(() => {
   }, [handleUnmount]);
 
   if (!wallet) {
-    return <Navigate to={{ pathname: '/', hash: 'sign-up' }} screenPath={ScreenPaths.GENERATION} />;
+    return <Navigate to={{ pathname: '/', hash: 'sign-up' }} screenPath={RouteUrls.SignUp} />;
   }
 
   return (