+++ src/pages/sign-out-confirm/sign-out-confirm.tsx
@@ -1,26 +1,26 @@
 import { useChangeScreen } from '@common/hooks/use-change-screen';
 import { useDrawers } from '@common/hooks/use-drawers';
 import { useWallet } from '@common/hooks/use-wallet';
-import { ScreenPaths } from '@common/types';
+import { RouteUrls } from '@routes/route-urls';
 import React from 'react';
 
 import { SignOutConfirmLayout } from './sign-out-confirm-layout';
 
 export const SignOutConfirmDrawer = () => {
-  const { doSignOut } = useWallet();
+  const { signOut } = useWallet();
   const changeScreen = useChangeScreen();
   const { setShowSignOut } = useDrawers();
 
   return (
     <SignOutConfirmLayout
       onUserDeleteWallet={async () => {
-        await doSignOut();
+        await signOut();
         setShowSignOut(false);
-        changeScreen(ScreenPaths.INSTALLED);
+        changeScreen(RouteUrls.Installed);
       }}
       onUserSafelyReturnToHomepage={() => {
         setShowSignOut(false);
-        changeScreen(ScreenPaths.HOME);
+        changeScreen(RouteUrls.Home);
       }}
     />
   );