+++ src/pages/signed-out/signed-out-view.tsx
@@ -4,12 +4,12 @@ import { Box, Button } from '@stacks/ui';
 import { PopupContainer } from '@components/popup/container';
 import { Text } from '@components/typography';
 
-import { ScreenPaths } from '@common/types';
+import { RouteUrls } from '@routes/route-urls';
 import { useChangeScreen } from '@common/hooks/use-change-screen';
 import { Header } from '@components/header';
 
 export const SignedOut = memo(() => {
-  const doChangeScreen = useChangeScreen();
+  const changeScreen = useChangeScreen();
   return (
     <PopupContainer header={<Header hideActions />}>
       <Box width="100%" mt="extra-loose" textAlign="center">
@@ -19,7 +19,7 @@ export const SignedOut = memo(() => {
         <Button
           my="extra-loose"
           onClick={() => {
-            doChangeScreen(ScreenPaths.INSTALLED);
+            changeScreen(RouteUrls.Installed);
           }}
         >
           Get started