+++ src/pages/unlock.tsx
@@ -7,24 +7,32 @@ import { ErrorLabel } from '@components/error-label';
 import { Header } from '@components/header';
 import { SignOutConfirmDrawer } from '@pages/sign-out-confirm/sign-out-confirm';
 import { useDrawers } from '@common/hooks/use-drawers';
+import { useAnalytics } from '@common/hooks/analytics/use-analytics';
 
 export const Unlock: React.FC = () => {
-  const { doUnlockWallet } = useWallet();
+  const { unlockWallet } = useWallet();
   const [loading, setLoading] = useState(false);
   const [password, setPassword] = useState('');
   const [error, setError] = useState('');
   const { showSignOut } = useDrawers();
+  const analytics = useAnalytics();
 
   const submit = useCallback(async () => {
+    const startUnlockTimeMs = performance.now();
+    void analytics.track('start_unlock');
     setLoading(true);
     setError('');
     try {
-      await doUnlockWallet(password);
+      await unlockWallet(password);
     } catch (error) {
       setError('The password you entered is invalid.');
     }
     setLoading(false);
-  }, [doUnlockWallet, password]);
+    const unlockSuccessTimeMs = performance.now();
+    void analytics.track('complete_unlock', {
+      durationMs: unlockSuccessTimeMs - startUnlockTimeMs,
+    });
+  }, [analytics, unlockWallet, password]);
 
   return (
     <>