+++ tests/integration/utils.ts
@@ -11,8 +11,12 @@ import { SettingsSelectors } from '@tests/integration/settings.selectors';
 
 const makeTmpDir = promisify(mkdtemp);
 
-export function createTestSelector(name: string) {
+export function createTestSelector<T extends string>(name: T): `[data-testid="${T}"]` {
   return `[data-testid="${name}"]`;
+}
+
+export function getCurrentTestName() {
+  return expect.getState().currentTestName.replaceAll(' ', '-');
 }
 
 export function randomString(len: number) {